ninjadsa/requirements.txt
streamlit==1.36.0
google-genai>=1.0.0